+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Started maintaining a changelog
+* `Config::comment_defaults` allows setting `//@` comments for all tests
 
 ### Fixed
 
@@ -20,5 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 * `$DIR` and `RUSTLIB` replacements
+* `Config::edition` (replaced by `config.comment_defaults.base().edition`)
 
 ## [0.21.2] - 2023-09-27